Enhance printing output support.
This patch is a major upgrade on the output formatting options
available for clidesc. It allows the display of enumerated lists,
formatting of list elements, optionally display element key, and
per element output formatting.

This is still in "preview" and might change without notice, but is
is a `good enough state` to be useful.