Fix tests for Python 3.10.
Python 3.10 is expected to be released in the last quarter of 2021,
and `argparse` seems to have home output changes that break `clidesc`
tests.

This PR introduces a step `Then output contains`, which uses a contex
text to provide a list of lines that should be found on the console
output. Each line is searched independently, and the order in which
they appear is not relevant.

Python 3.10 is not enabled by default in tox, and are not part of
upstream CI. The tests for 3.10 must be executed with `tox -e py310`.